+++ api/app/Http/Controllers/OrganizationController.php
@@ -102,6 +102,12 @@ class OrganizationController extends Controller
     public function destroy(OrganizationRequest $request, $id)
     {
         $o = Organization::findOrFail($id);
+
+        // Make sure that the organization does not have any facilities.
+        if ($o->facilities()->first()) {
+            abort(400);
+        }
+
         $deletedOrg = $this->toCcArray($o->toArray());
         $o->delete();
         return $deletedOrg;
